6.16 I'm going to be now trying to add some sort of config option to this color picker to demonstrate how we actually pass that in. Very similar to our text input demo. So let's have a look at what options we have to play with.
 So for example, let's have a look at making a width method on the color picker. So we can pass in something like width, and then 100, and then that changes it from 320 to be 100 instead. So how we do this is we make our public method on our field instance. So public function width, we pass in our int of width, we're going to be returning this current type of this should probably be static, so that if someone inherits this color picker, then we're going to be returning the correct child class instead of the current class.
 And we're going to be setting a property to be the width, so this width equals width. Now one oversight I did make was this should probably also accept a closure, so that we
VM1592:2 can evaluate it. And potentially the default should be null, and then only if we pass in a width should we actually pass that into the JavaScript library else it should dictate how wide it should be. So let's make our property so protected int closure or null width, and the default for this is going to be null. And then we need our getter, so public function get width int closure null.